+++ datasets/fixmatch_cifar.py
@@ -40,12 +40,12 @@ def get_cifar10(args, root, labels_to_tensor=False):
     train_labeled_dataset = CIFAR10SSL(
         root, train_labeled_idxs, train=True,
         transform=transform_labeled,
-        target_transform=lambda x: torch.tensor(x, dtype=torch.int8) if labels_to_tensor else None)
+        target_transform=lambda x: torch.tensor(x).type(torch.LongTensor) if labels_to_tensor else None)
 
     train_unlabeled_dataset = CIFAR10SSL(
         root, train_unlabeled_idxs, train=True,
         transform=TransformFixMatch(mean=cifar10_mean, std=cifar10_std),
-        target_transform=lambda x: torch.tensor(x, dtype=torch.int8) if labels_to_tensor else None)
+        target_transform=lambda x: torch.tensor(x).type(torch.LongTensor) if labels_to_tensor else None)
 
     test_dataset = datasets.CIFAR10(
         root, train=False, transform=transform_val, download=False)
@@ -156,12 +156,12 @@ def get_cifar100(args, root, labels_to_tensor=False):
     train_labeled_dataset = CIFAR100SSL(
         root, train_labeled_idxs, train=True,
         transform=transform_labeled,
-        target_transform=lambda x: torch.tensor(x, dtype=torch.int8) if labels_to_tensor else None)
+        target_transform=lambda x: torch.tensor(x).type(torch.LongTensor) if labels_to_tensor else None)
 
     train_unlabeled_dataset = CIFAR100SSL(
         root, train_unlabeled_idxs, train=True,
         transform=TransformFixMatch(mean=cifar100_mean, std=cifar100_std),
-        target_transform=lambda x: torch.tensor(x, dtype=torch.int8) if labels_to_tensor else None)
+        target_transform=lambda x: torch.tensor(x).type(torch.LongTensor) if labels_to_tensor else None)
 
     test_dataset = datasets.CIFAR100(
         root, train=False, transform=transform_val, download=False)
